Establish a default report dir, and then reports grouped by project title.
i.e. if you are using an inputfile and the title of that project
is "My Project", there will be created a dir "My Project" in the
default report dir, and any reports for it will be put there.

Can be overwritten by --output arg; specify an abs value there
and it will always priorities that.