Remove second GUI version.
I submitted a second version of the GUI here:
 cd5d5d61423eef845530183798b1eb5c7663c0e5

In this one, the form submits to itself, so that the php and HTML
are all on the same page. I was doing this so that I could update
the div that holds the result of the report once the php completes.
However, it seems really convoluted and messy. Just going to stick
to having index.html which will open to upload.php.