Add encoding in docx --> txt case, in case file has Russian chars
In last commit, I added functionality to write docx files to txt
files to back them up.

I discovered that if there's Cyrillic chars in that, the script
will then fail - not on the write - but later on the read when it's
trying to get the word count on the txt file that was generated.

If I added encoding='utf-8' to the read, that was not sufficient,
nor is it sufficient to add the encoding on the write alone. Need
to add the encoding on both the write of the txt file, as well
as the later read.